chore: Move catalogPlugin to catalog-backend
