+++ plugins/catalog-backend/package.json
@@ -34,6 +34,8 @@
     "clean": "backstage-cli package clean"
   },
   "dependencies": {
+    "@backstage/backend-app-api": "^0.0.0",
+    "@backstage/backend-plugin-api": "^0.0.0",
     "@backstage/plugin-catalog-node": "^0.0.0",
     "@backstage/backend-common": "^0.14.1-next.2",
     "@backstage/catalog-client": "^1.0.4-next.1",
